requirements.txt
Django==1.9.2
django-filter==0.12.0
djangorestframework==3.3.2
Markdown==2.6.5
wheel==0.24.0
